requirements.txt
GitPython ~= 3.1
packaging >= 23.0
semver ~= 2.13
setuptools >= 67.4
